refactor(isa): split IProcessor into agnostic base and mips::IMipsProcessor
Separate the ISA-agnostic processor contract from MIPS-specific architectural
state so a RISC-V core can reuse the same interface, TUI, and Qt/QML
visualizers without inheriting MIPS coprocessor internals.

- include/isa/processor.h: new isa::IProcessor holding StepResult,
  StageSnapshot, PipelineState and the generic accessors (load_program, step,
  run, reset, pc, regs, mem, cycle_count, pipeline_state).
- mips/processor.h: mips::IMipsProcessor : isa::IProcessor adds the MIPS-only
  bits (last_control(), cp0(), hi()/lo(), set_hi/lo). Control (with the
  MIPS-only reg_dst) and derive_control stay in mips. mips::IProcessor /
  StepResult / PipelineState / StageSnapshot are kept as `using` aliases so
  generic callers are untouched.

SingleCycleCpu and PipelinedCpu now derive from IMipsProcessor. GdbStub and
nsc/ui.cpp's owned cpu pointer are retyped to IMipsProcessor (they use cp0/
hi/lo/last_control); SimulatorController and the render_* helpers stay on the
agnostic IProcessor. No behavior change; full debug build (TUI + Qt Widgets +
QML) and all 22 tests green.
